+++ src/Tools.php
@@ -22,6 +22,7 @@ use spicyweb\oddsandends\fields\DisabledNumber as DisabledNumberField;
 use spicyweb\oddsandends\fields\DisabledPlainText as DisabledPlainTextField;
 use spicyweb\oddsandends\fields\EntriesSearch as EntriesSearchField;
 use spicyweb\oddsandends\fields\Grid as GridField;
+use spicyweb\oddsandends\fields\ProductsSearch as ProductsSearchField;
 use spicyweb\oddsandends\fields\Width as WidthField;
 use spicyweb\oddsandends\models\Settings;
 use spicyweb\oddsandends\widgets\RollYourOwn as RollYourOwnWidget;
@@ -73,6 +74,11 @@ class Tools extends Plugin
                 $event->types[] = WidthField::class;
                 $event->types[] = AncestorsField::class;
                 $event->types[] = GridField::class;
+
+                $pluginsService = Craft::$app->getPlugins();
+                if ($pluginsService->isPluginInstalled('commerce') && $pluginsService->isPluginEnabled('commerce')) {
+                    $event->types[] = ProductsSearchField::class;
+                }
             }
         );