+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Added
+- Added `spicyweb\oddsandends\fields\ProductsSearch` - Commerce Products (Search) field type
+
 ## 4.0.0 - 2022-10-19
 
 ### Added